pong

+++ src/pixie/iomem.h
@@ -72,7 +72,7 @@ struct s_Color {
 
 typedef struct s_Map Map;
 struct s_Map {
-	uint16_t addr;	///< Address of the map data
+	uint32_t addr;	///< Address of the map data
 	uint16_t aux;	///< Address of the auxillary map data (if equal to addr then not used)
 	uint16_t w,h;	///< Width/Height of the layer map (in tiles)
 	int16_t x,y;	///< Signed screen offset in pixels
@@ -80,7 +80,7 @@ struct s_Map {
 
 typedef struct s_Tiles Tiles;
 struct s_Tiles {
-	uint16_t addr;		///< Address of the tile data
+	uint32_t addr;		///< Address of the tile data
 	uint16_t w,h;		///< Width / Height of one tile in pixels
 	uint16_t row_bytes;	///< Number of bytes to encode one row of a tile
 	uint8_t bits;		///< Number of bits per pixel (1, 2, 4)
@@ -91,7 +91,7 @@ typedef struct s_Layer Layer;
 struct s_Layer {
 	Map map;
 	Tiles tiles;
-	uint16_t palette;	///< Address of the palette data
+	uint32_t palette;	///< Address of the palette data
 	uint8_t border;		///< How tiles outside the map are handled.  x(high nibble) y(low nibble)
 	uint8_t flags;		
 };
